requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
apt-wrapper==1.18
asgiref==3.7.2
async-timeout==4.0.3
attrs==24.2.0
Babel==2.8.0
bcc==0.1.10
billiard==4.2.0
blinker==1.4
Brlapi==0.8.3
celery==5.3.6
certifi==2024.8.30
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
chrome-gnome-shell==0.0.0
ci-info==0.3.0
cliapp==1.20180812.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
configobj==5.0.8
configparser==7.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.4
distlib==0.3.8
distro==1.9.0
dj-database-url==2.2.0
Django==4.2.6
django-cors-headers==4.3.0
django-environ==0.9.0
django-filter==2.4.0
django-guardian==2.4.0
djangorestframework==3.14.0
easyocr==1.7.1
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.2.2
fastapi==0.112.1
filelock==3.15.4
fitz==0.0.1.dev2
frozenlist==1.4.1
fsspec==2024.9.0
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
hidpidaemon==18.4.6
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.2
idna==3.8
imageio==2.35.1
importlib-metadata==4.6.4
isodate==0.6.1
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
kernelstub==3.1.4
keyring==23.5.0
kombu==5.3.5
langchain==0.2.14
langchain-core==0.2.33
langchain-openai==0.1.22
langchain-text-splitters==0.2.2
langsmith==0.1.100
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
looseversion==1.3.0
louis==1.3
lxml==5.3.0
macaroonbakery==1.3.1
Markdown==3.3.6
MarkupSafe==2.1.5
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.5
netaddr==0.8.0
netifaces==0.11.0
networkx==3.3
nibabel==5.2.1
ninja==1.11.1.1
nipype==1.8.6
numpy==2.1.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
openai==1.42.0
opencv-python-headless==4.10.0.84
openpyxl==3.0.10
orjson==3.10.7
packaging==24.1
pandas==1.5.2
pathlib==1.0.1
pillow==10.4.0
pipdeptree==2.13.2
pipenv==11.9.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
protobuf==3.12.4
prov==2.0.1
psycopg==3.1.17
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycairo==1.20.1
pyclipper==1.3.0.post5
pycparser==2.21
pycups==2.0.1
pydantic==2.8.2
pydantic_core==2.20.1
pydbus==0.6.0
pydot==3.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==3.1.2
PyPDF2==3.0.1
pyRFC3339==1.1
pytesseract==0.3.13
python-apt==2.4.0+ubuntu3
python-bidi==0.6.0
python-dateutil==2.8.2
python-debian==0.1.49
python-dotenv==1.0.1
python-gnupg==0.4.8
python-systemd==0.0.9
python-telegram-bot==21.5
python-xlib==0.29
pythondialog==3.5.3
pytz==2022.1
pywal==3.3.0
pyxdg==0.27
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.7.24
repolib==1.5.1
repoman==1.4.0
requests==2.32.3
scikit-image==0.24.0
scipy==1.14.1
SecretStorage==3.3.1
sessioninstaller==0.0.0
shapely==2.0.6
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
sqlparse==0.4.4
starlette==0.38.2
sympy==1.13.2
systemd-python==234
tenacity==8.5.0
tifffile==2024.8.30
tiktoken==0.7.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
traitlets==5.14.1
traits==6.3.2
traittypes==0.2.1
triton==3.0.0
ttystatus==0.38
tufw==1.1.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
uritemplate==3.0.1
urllib3==1.26.5
uvicorn==0.30.6
vine==5.1.0
virtualenv==20.25.0
virtualenv-clone==0.3.0
wadllib==1.3.6
wcwidth==0.2.13
xdg==5
xkit==0.0.0
XlsxWriter==3.0.8
yarl==1.9.4
zipp==3.17.0
